V3: Added the ability to disable the Header
The option "loa_title" (already implemented) controls this behaviour.

Note: I want to retain the same behaviour as the previous version: by
default, if this option is set to nil, a default Header Title is set.
To implement this V3, I make a difference between nil (not set) and empty
(""). If the title is empty, we disable the Header. If it is not set,
we replace by a default.