Added the ability to define and load acronyms from external files.
Those external files must have a Markdown format, but only
the YAML metadata will be used. They can be specified with the
'acronyms.fromfile' option in the main Markdown document.
This option can be either: a single string (for a single file),
or a list of strings (which corresponds to a list of files).