Refactored tests for better readability.
Moved all the calls to 'testthat::test_that' in a separate file ('test_rmarkdown.R') and replaced them to 'testthat::test_file' in the main
'testthat.R' script.
In practice, this means that testthat creates a nice report that is easier to read.
Also, configured 'rmarkdown' to be quiet, which reduces the clutter in the output. (We do not care what files are created anyway, since they are temporary and destroyed after the test ends)