Allow using several pandoc args + better doc
Previous version of the 'add_filter()' function would return the args as a vector. This made it easy to use with 'pandoc_args: !expr add_filter()', but this syntax does not allow adding any other arguments to 'pandoc_args'.
The new version returns the args as a single string, which means it can be used as previously with 'pandoc_args: !expr add_filter()', but also in a list. For example:
pandoc_args:
  - !expr add_filter()
  - --wrap=preserve

Also improved the documentation of 'add_filter', and particularly added more examples of how to use it in the YAML metadata.