Made tests easier to write by preserving line wraps
Previous version let the automatic behaviour of Pandoc handle wraps, which, by default, break lines after a given number of columns (usually, 72).
This would make it harder to write the "expected" files for each test, since we compare the files line-by-line. We had to determine where Pandoc would break lines, after replacing the acronyms, to correctly format the expected file.

The new version preserves the line breaks, which means that Pandoc will only break a line where there is a line break in the original ("input.Rmd") file.
We can therefore focus on comparing the semantic content of the files, and spend less time on where the lines break.