Refactor 'insert_beginning' into 'insert_loa'
Previously, the option 'insert_beginning' could be used to either insert the List Of Acronyms (LoA) at the beginning of the document, or to not insert it.
The name was somewhat misleading, and offered very limited behaviour.

The new option 'insert_loa' replaces 'insert_beginning', and allows the following values:
* false -> the LoA is not automatically inserted (but can still be inserted manually with '\printacronyms' ; it is equivalent to 'insert_beginning: false')
* 'beginning' -> the LoA is automatically inserted at the beginning (equivalent to 'insert_beginning: true')
* 'end': -> the LoA is automatically inserted at the end (a new behaviour that was not available before)