Improved testing workflow
The 'R CMD CHECK' step was moved to the end of the workflow, after we run the automated tests.
Indeed, this step runs the automated tests itself, but with a far less readable report.
If the tests should fail, it would be better that this happen in a dedicated step for better readability.