Fixes for pandoc 2.17.0.1
The main breaking change for acronymsdown is that the MetaList values do not have a '.tag' attribute anymore.
We used this attribute to see if a value was a list or an inline, on 2 occasions:
  - when parsing the acronyms keys ;
  - when parsing the 'fromfile' field.

This commit adds a new helper function 'isMetaList' that performs this type checks differently for pandoc >= 2.17 and pandoc < 2.17.

* Acronyms were no longer parsed, because of the failed check. We replaced it by 'isMetaList'.

* Multiple external files were treated as a single file, because of the same failed check. We replaced it by 'isMetaList'.

* The 'Doc' function was removed, which we previously used to automatically insert the LoA ('appendLoa(doc)'). The 'Pandoc' filter should work in the exact same way, even in older versions.