Stop on failed tests
The previous version of 'make tests' did not stop when one of the tests fail (because testthat was simply not returning a non-zero code).
Only the 'R CMD check' (in 'make check') would fail ; however the reporter used in 'check' is less readable.

This new version makes testthat abort (thus returning a non-zero code) if at least one test has failed. All tests still are still ran first.

The advantage is that, in a testing workflow, the following steps will not be executed.

Also, removed the 'testthat.R' file since it was not really useful. Because it relied on 'test_rmarkdown.R' and this latter was in the same dir, in some contexts we would run tests twice (especially in 'make check'), which is useless and a waste of time.
We now directly run 'test_rmarkdown.R' from 'make tests'.