Added a custom output format
The new 'acronymsdown::md_format' is used for facilitating the automated tests.
It automatically sets the same output parameters that were previously manually defined in the tests 'input.Rmd' files:
* the acronymsdown Lua Filter
* the '--wrap=preserve' option
* the (pandoc) 'markdown' variant.
In addition, it adds a new parameter to ensure that headings follow the ATX syntax. This parameter depends on the Pandoc version, and thus could not be manually defined in the 'input.Rmd' files previously.