Updated the tests to use the custom output format
This custom format greatly simplifies the test files, and allows to focus on the behaviour we want to test itself.
Also, we can change a parameter in the custom format itself, instead of altering all test files.
Since the output format adds a parameter to ensure that ATX headings are produced, even on older versions of pandoc, this allows testing acronymsdown with pandoc < 2.11.2, which was not feasible previously (the tests would fail because of the headings).