Fixed issue when importing `acronyms.lua` and using Quarto
This commit fixes #7 in which the `acronyms.lua` could not be imported
(through `require`) when using the `parse-acronyms.lua` script with
Quarto.
Only the `acronyms.lua` file seemed to cause problems; coincidentally,
it is the only file that does not `return` a value, i.e., the `Acronyms`
table was loaded as a global value.
Perhaps something in the way Quarto loads the filters prevents this kind
of requires.

This commit:

* Adds the `Acronym` table as an element of the `Acronyms` table so that
  it can be accessed easily, while not being a global value.
* Returns the `Acronyms` table at the end of the `acronyms.lua` file.
* Explicitly affect the return of the `require('acronyms.lua')` to a
  `local Acronyms` variable.

It therefore offers the same features as previously, but using a
"cleaner" import.