Fixed the pkgdown workflow
The pkgdown workflow used old versions (v1) of actions from r-lib,
such as setup-pandoc, setup-r, and setup-r-dependencies.
These versions seemed not to work anymore, in part because they
rely on Node.js 12 which was deprecated by GitHub, and in part
because they tried to download older versions of R.

Modified the `r-setup-dependencies` configuration to follow
the example from r-lib: we need `any::pkgdown` to install pkgdown,
and add `local::.` to install our own library (which is then required
by pkgdown when building the website).