Added mention to the new acronyms extension for Quarto
Several Quarto users were interested in the Acronymsdown package,
but wanted to use it with Quarto rather than RMarkdown. While this
was already feasible (by copying and pasting the Lua files), this
was not an easy nor automated process. Providing a Quarto extension
makes it easier for these users, and will allow us to implement
new features targeting Quarto documents (such as shortcodes).

This commit closes #6.